requirements.txt
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cryptography==44.0.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
Flask==3.0.3
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==2.1.5
numpy==1.24.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
pandas==2.0.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pyasn1==0.4.5
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyMySQL==1.1.1
pypdfium2==4.30.1
pysmb==1.2.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.33.0
Werkzeug==3.0.6
zipp==3.20.2
